requirements.txt
black==20.8b1
tensorflow==2.4.0
pylint==2.6.0
sklearn==0.0
pandas==1.2.0
matplotlib==3.3.3
pygit2==1.4.0
PyYAML==5.3.1
